DBZ-4910 Run Cassandra in container
To be able to run tests with Java 11+, we have to run Cassandra in
a container as Cassandra 3 doesn't work with Java 11+.

As the tests need to access Cassandra files, dedicated directory is
created in `target` dir, container bind mounts this directory and
Cassandra is configured to use this mounted dir for storing data.
As this directory has to be writable by Cassandra user, which may not
exists on test machine, folder has 777 rights.

Container with Cassandra is started in `@BeforeClass` method and
torn down in `@AfterClass` class method.

Containers and their images are created on the fly and removed after the
test. Path to the Dockerfile is specified by `docker.dir` system
property.

Because data written in `target` dir by Cassadra from container are
written by different user than one running the test, it cannot be
removed. Removing the data on Cassandra exit doesn't seem very realible,
as Cassandra may still write some data during exit. The most simple and
realibale way seems to by be to run clean-up container in `@AfterClass`
class method once Cassandra container is stopped, which bind mounts the
same direcotry and removes all the Cassandra files. This container is
defined direcly in the code and there's no Dockerfile for this
container.

Also adjust existing tests not to access Cassandra commit log direcly.
As Cassandra runs now in a container, there's no need to insert records
direcly into the log and can be inserted normally using the clinet.
Accessing the commit log from test freezes the test, so also one assert
is removed.