DBZ-4910 Use Java 11+
Remove constraints on running with Java 8 and enable to run the tests
with Java 17. Connector itself is build with Java 11, tests run with
Java 17.

There is still one error in the logs for Cassandrs 3:

    java.lang.NoSuchMethodError: 'sun.misc.Cleaner sun.nio.ch.DirectBuffer.cleaner()'

This is due to the fact that `Cleaner` was moved into `jdk.internal.ref`
package [1]. However, this error doesn't affect the tests and can be
ignored.

[1] https://github.com/openjdk/jdk/commit/5765b5b4f03fa4c4d12f636424ac49e264212e1c