DBZ-7554 AvroConverter fails to parse events
Adjust schema names for cells to be unique. Avro uses these names to
identify sub schema declarations and will not redefine a sub schema that
was already detected. In this case, if we have cells with multiple
types, debezium will fail to serialize the event, since the stored avro
schema does not match the cells.